+++ theme/template/archive.php
@@ -6,9 +6,7 @@
  *
  * @package _s
  */
-
 get_header(); ?>
-
 	<div id="primary" class="content-area">
 		<main id="main" class="site-main">