+++ theme/template/comments.php
@@ -1,1 +1,1 @@
-Not implemented
+Comments not implemented.